Allow models to intuit the providers their associations belong to if the name of the provider is the same as the name of the association. E.g.:
	function System(data) {
		this.hasMany('sensors');
	}

No provider needs to be declared as long as the name of the provider is "Sensor". 

Add the ability to declare foreign keys (though these aren't yet used). 

Handle $saves where multiple foreign keys are returned and need to be looked up, possibly calling multiple async $http requests to the API for these associations to populate. 

Declare belongs to associations as enumerable, so they now show up on the instances.